+++ src/spotted/handlers/sban.py
@@ -16,7 +16,7 @@ async def sban_cmd(update: Update, context: CallbackContext):
     """
     info = EventInfo.from_message(update, context)
     if context.args is None or len(context.args) == 0:  # if no args have been passed
-        banned_users = "\n".join(str(user.user_id) for user in User.banned_users())
+        banned_users = "\n".join(f"{user.user_id} ({user.ban_date:%d/%m/%Y %H:%M})" for user in User.banned_users())
         banned_users = "Nessuno" if len(banned_users) == 0 else f"{banned_users}"
         text = f"[uso]: /sban <user_id1> [...user_id2]\nGli utenti attualmente bannati sono:\n{banned_users}"
         await info.bot.send_message(chat_id=info.chat_id, text=text)